move `useCapabilitiesSafe` to `internal/hooks`
