+++ src/internal/hooks/useCapabilitiesSafe.test.ts
