+++ src/internal/hooks/useCapabilitiesSafe.ts
@@ -1,7 +1,7 @@
 import type { WalletCapabilities } from 'viem';
 import { useAccount } from 'wagmi';
 import { useCapabilities } from 'wagmi/experimental';
-import type { UseCapabilitiesSafeParams } from './types';
+import type { UseCapabilitiesSafeParams } from '../../types';
 
 export function useCapabilitiesSafe({
   chainId,
